new types and cleaning

+++ src/tx-handlers/handleMsgDeleteCollection.ts
@@ -1,28 +1,24 @@
-import { Collection, DbStatus, DocsCache } from "bitbadgesjs-utils"
-import { MessageMsgDeleteCollection } from "bitbadgesjs-transactions"
-import { BALANCES_DB, CLAIMS_DB, METADATA_DB, fetchDocsForCacheIfEmpty } from "../db/db"
+import { MsgDeleteCollection } from "bitbadgesjs-transactions"
+import { DocsCache, StatusDoc } from "bitbadgesjs-utils"
+import { BALANCES_DB, CLAIMS_DB } from "../db/db"
 
-import nano from "nano"
+import { fetchDocsForCacheIfEmpty } from "src/db/cache"
 import { handleNewAccountByAddress } from "./handleNewAccount"
 
-export const handleMsgDeleteCollection = async (msg: MessageMsgDeleteCollection, status: DbStatus, docs: DocsCache): Promise<void> => {
-  await fetchDocsForCacheIfEmpty(docs, [msg.creator], [msg.collectionId], [], [], []);
+export const handleMsgDeleteCollection = async (msg: MsgDeleteCollection<bigint>, status: StatusDoc<bigint>, docs: DocsCache): Promise<void> => {
+  await fetchDocsForCacheIfEmpty(docs, [msg.creator], [msg.collectionId], [], []);
   await handleNewAccountByAddress(msg.creator, docs);
 
-
   //Safe to cast because MsgDeleteCollection can only be called if the collection exists
-  const collectionDoc = docs.collections[msg.collectionId.toString()] as Collection & nano.DocumentGetResponse;
+  const collectionDoc = docs.collections[msg.collectionId.toString()];
+  if (!collectionDoc) throw new Error(`Collection ${msg.collectionId.toString()} does not exist`);
   collectionDoc._deleted = true;
 
   //Delete all relevant docs from DB
-  const allMetadataDocs = await METADATA_DB.partitionedList(`${msg.collectionId.toString()}`);
   const allBalancesDocs = await BALANCES_DB.partitionedList(`${msg.collectionId.toString()}`);
   const allClaimsDocs = await CLAIMS_DB.partitionedList(`${msg.collectionId.toString()}`);
 
   const promises = [];
-  for (const doc of allMetadataDocs.rows) {
-    promises.push(METADATA_DB.destroy(doc.id, doc.value.rev));
-  }
   for (const doc of allBalancesDocs.rows) {
     promises.push(BALANCES_DB.destroy(doc.id, doc.value.rev));
   }
@@ -31,13 +27,6 @@ export const handleMsgDeleteCollection = async (msg: MessageMsgDeleteCollection,
   }
 
   await Promise.all(promises);
-
-  //Delete all metadata docs from docs object (safe to do because no TXs will be processed after this delete one)
-  for (const key of Object.keys(docs.metadata)) {
-    if (key.split(':')[0] === `${msg.collectionId}`) {
-      delete docs.metadata[key];
-    }
-  }
 
   for (const key of Object.keys(docs.balances)) {
     if (key.split(':')[0] === `${msg.collectionId}`) {
